I believe I have what was asked of me
I defined a max distance "testDistance" that updates for every test case. I then run findPerms. In findPerms I used the algo we talked about in class but for when k == n I calculated a new current distance for this set of permutations. On the first run curdistance will be less than testDistance so it will display the perm and then calc the distances for the permutation then is will update testDistance. The program will then return to find perms and get a new permutation. The program will then find the total distance for the new perm and test it against the updated testDistance. If it is not less than it will not display the perm or calculate any distances.